+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added: `declaration-nested-properties` rule.
 - Added: `declaration-nested-properties-no-divided-groups` rule.
 - Added: `ignore: "local"|"global"` to the `dollar-variable-pattern` rule.
+- Added: `dollar-variable-empty-line-before` rule.
 
 # 1.3.4